Start making search bar actually usable
